Refactor Headers class
