+++ src/Headers.php
@@ -11,6 +11,7 @@ declare(strict_types=1);
 
 namespace HttpClient;
 
+use HttpClient\Header\HeaderFilter;
 use HttpClient\Util\HttpSecurity;
 use Psr\Http\Message\UriInterface;
 
@@ -39,8 +40,8 @@ final class Headers
      */
     public function __construct(UriInterface $uri, array $headers = [])
     {
-        [$this->headerNames, $headers] = self::filterHeaders($headers);
-        $this->assertHeaders($headers);
+        [$this->headerNames, $headers] = HeadersFilter::filterHeaders($headers);
+        HeadersFilter::assertHeaders($headers);
         $this->headers = $headers;
         $this->uri = $uri;
     }
@@ -71,31 +72,7 @@ final class Headers
         
         return $headers;
     }
-    
-    /**
-     * @param $name
-     * @param $value
-     * @return array
-     * @throws \InvalidArgumentException
-     */
-    public static function assertValidHeader($name, $value): array
-    {
-        if (is_string($value)) {
-            $value = [$value];
-        }
-        
-        if (!is_array($value) || !self::arrayContainsOnlyStrings($value)) {
-            throw new \InvalidArgumentException(
-                'Invalid header value; must be a string or array of strings'
-            );
-        }
-        
-        HttpSecurity::assertValidName($name);
-        self::assertValidHeaderValue($value);
-        
-        return [$name, $value];
-    }
-    
+
     /**
      * @param $name
      * @param $value
@@ -158,8 +135,8 @@ final class Headers
      */
     public function setHeaders(array $headers)
     {
-        [$this->headerNames, $headers] = self::filterHeaders($headers);
-        $this->assertHeaders($headers);
+        [$this->headerNames, $headers] = HeadersFilter::filterHeaders($headers);
+        HeadersFilter::assertHeaders($headers);
         
         $this->headers = $headers;
         
@@ -252,100 +229,5 @@ final class Headers
         unset($this->headers[$header], $this->headerNames[$normalizedHeader]);
         
         return $this;
-    }
-    
-    /**
-     * @param array $headers
-     * @throws \InvalidArgumentException
-     */
-    private function assertHeaders(array $headers): void
-    {
-        foreach ($headers as $name => $headerValues) {
-            HttpSecurity::assertValidName($name);
-            array_walk($headerValues, __NAMESPACE__ . '\Util\HttpSecurity::assertValid');
-        }
-    }
-    
-    /**
-     * @param array $array
-     * @return mixed
-     */
-    private static function arrayContainsOnlyStrings(array $array)
-    {
-        return array_reduce($array, [__CLASS__, 'filterStringValue'], true);
-    }
-    
-    /**
-     * @param array $values
-     */
-    private static function assertValidHeaderValue(array $values): void
-    {
-        array_walk($values, __NAMESPACE__ . '\Util\HttpSecurity::assertValid');
-    }
-    
-    /**
-     * @param $value
-     * @throws \InvalidArgumentException
-     */
-    private static function assertHeaderValueType($value): void
-    {
-        if (!is_array($value) && !is_string($value) && !is_numeric($value)) {
-            throw new \InvalidArgumentException(sprintf(
-                'Invalid header value type; expected number, string, or array; received %s',
-                (is_object($value) ? get_class($value) : gettype($value))
-            ));
-        }
-        
-        if (is_array($value)) {
-            array_walk($value, function ($item) {
-                if (!is_string($item) && !is_numeric($item)) {
-                    throw new \InvalidArgumentException(sprintf(
-                        'Invalid header value type; expected number, string, or array; received %s',
-                        (is_object($item) ? get_class($item) : gettype($item))
-                    ));
-                }
-            });
-        }
-    }
-    
-    /**
-     * @param array $originalHeaders
-     * @return array
-     * @throws \InvalidArgumentException
-     */
-    private static function filterHeaders(array $originalHeaders): array
-    {
-        $headerNames = $headers = [];
-        foreach ($originalHeaders as $header => $value) {
-            if (!is_string($header)) {
-                throw new \InvalidArgumentException(sprintf(
-                    'Invalid header name; expected non-empty string, received %s',
-                    gettype($header)
-                ));
-            }
-            
-            self::assertHeaderValueType($value);
-            
-            $value = (array)$value;
-            
-            $headerNames[self::normalizeHeader($header)] = $header;
-            $headers[$header] = $value;
-        }
-        
-        return [$headerNames, $headers];
-    }
-    
-    /**
-     * @param $carry
-     * @param $item
-     * @return string|bool
-     */
-    private static function filterStringValue($carry, $item)
-    {
-        if (!is_string($item)) {
-            return false;
-        }
-        
-        return $carry;
     }
 }
